Repair 'True' and 'False' use with user config
This issue only affected CPython 2.7 where True and False are builtins.
The _safe_eval() code was excluding True and False from its _safe_builtins
list which thus made them unavailable on CPython 2.7.

The solution is to add True and False to the _safe_builtins list.

Unit tests are added that expose this problem.