simulate() can suppress exceptions
N.B. this is a potentially breaking change due to:

- Component no longer has `self.tracemgr` attribute. Instead has
  `self.env.tracemgr`.
- Tracer's no longer implement the context manager protocol. simulate()
  now takes responsibility for closing the TraceManager instance.

Perhaps the most important behavior change is that simulate() now has a
`reraise` parameter that controls whether unhandled exceptions are
re-raised to the caller of simulate(). This parameter is True by default,
so single-simulation behavior is preserved.

The default behavior of simulate_many() and simulate_factors() changes.
When running multiple simulations, simulate() is called with
`reraise=False`. This means that the caller of
simulate_many()/simulate_factors() will always be returned a complete list
of result dicts. The onus is on the caller to inspect the 'sim.exception'
result dict item to determine whether a simulation failed with an unhandled
exception.

This changeset also greatly bolsters the unit tests for the
desmod.simulation model. Test coverage now approaches 90%.