simulate_factors() now has a factors parameter
N.B. This is a breaking change.

This replaces the 'sim.factors' config item with an explicit parameter to
simulate_factors().

The 'sim.factors' config item's only purpose was to inform
simulate_factors() of the list of factors to expand. The 'sim.factors'
config item was not used within the generated config dicts and thus never
manifested in the result dicts or result YAML file.