Add flush() to tracing subsystem
TraceManager.flush() is added as a public interface. Tracer.flush() is also
added.

The goal for these methods is to allow the user to force a flush of traced
data. Flushing periodically may be useful when simulations are observed in
realtime and/or when debugging.

LogTracer and VCDTracer both implement flush() by calling flush() on their
respective file objects.