Parse unit in SimEnvironment.time()
The unit string parameter to SimEnvironment.time() was assumed to be just
the unit component. The unit may now be a complete timescale string--i.e.
it may also include a magnitude with the unit.

This is more consistent with other time specifications, but remains
compatible with any existing valid uses of SimEnvironment.time().