Add "persist" option for tracers
The various tracers now have a 'sim.xxx.persist' configuration option. When
disabled, the file(s) associated with the tracer will be removed at end of
simulation when the TraceManager is closed.

This makes it possible for simulation artifacts to only persist long enough
for any post-simulation processing to take place. This can be useful when
the trace artifact is sufficiently large that the cost of storage is a
concern.