Support progress notification on spawned processes
When running multiple simulations with simulate_many(), the progressbar
mechanism only worked when the pool processes were started using 'fork',
but deadlocked when using 'spawn'. Since spawn is the only start method on
Windows, that meant that multi-simulation progress was not supported on
Windows.

To repair the problem, we launch our own worker processes instead of using
a multiprocessing.Pool. This allows us to instantiate and explicitly pass
multiprocessing.Queue objects in a manner that is spawn-compatible.

In addition to passing a queue for simulation progress, we now also use a
config/result queue pair to feed the worker processes and gather result
dictionaries.