Repair race between py.test and progress thread
In multi-factor simulations (i.e. when using simulate_many()), there is a
progress thread that consumes progress information from the various
simulation processes and prints a progress display to stderr. The progress
thread is a daemon thread, so the strategy was to not join with it and
instead allow it to be auto-terminated, if necessary, when the main thread
terminated.

However, this strategy is exposed to a race condition with py.test. The
problem is that py.test wraps sys.stderr and may close sys.stderr before
the progress thread (which uses sys.stderr) is done with it. In a regular
environment, sys.stderr is not explicitly closed and is thus always
available to the progress thread.

A two-fold strategy is employed to repair this race opportunity. First, we
now attempt to join the progress thread. This join is done with a short, 1
second timeout so as to not allow the main thread to get stuck indefinitely
if the progress thread is blocked waiting for the progress queue.

Second, we put some safety rails in _consume_multi_display_simple(). This
is the particular display function affected by py.test closing stderr. We
avoid extra isatty() calls (because those fail when the file is closed) as
well as testing fd.closed before writing to it.