Add sim.vcd.start_time and sim.vcd.stop_time
These new configuration values control the periods within the simulation
that VCD tracing is enabled/disabled. The VCDWriter.dump_on() and
dump_off() methods are used to implement this feature.

Note that various combinations of start_time and stop_time may be used to
determine the time-region(s) where VCD is enabled. For example, specifying
a stop_time less than the start_time will cause VCD to be enabled at the
beginning and end of simulation, but disabled in the middle of simulation.
Here is a pictoral illustration of the possible configurations:

    |vvvvvvvvvvvv|  neither start nor stop
    |----vvvv----|  start < stop
    |vvvv-----vvv|  stop < start
    |----vvvvvvvv|  start only
    |vvvvvvv-----|  stop only