meta.sim.index and meta.sim.special
A new 'meta' namespace is introduced. The 'sim.seq' config item, which is
set by simulate_many() is now called 'meta.sim.index'. The new namespace is
meant to avoid the problem of desmod modifying the user-configurable 'sim'
namespace. The new 'index' name is meant to be less obscure than 'seq'.

When using simulate_factors(), the "special" configuration items can
optionally be added to the config. The default is now 'meta.sim.special'
instead of 'sim.special' in order to keep the 'sim' namespace clean.