Add named configuration categories and doc strings
NamedManager.name() now optionally accepts category and doc parameters.
This new metadata is made available when iterating the named
configurations.

A new NamedConfig namedtuple is introduced to make it easier to manage the
now-larger set of items associated with each name.

NamedManager now uses iterator protocol. The iter() method is replaced with
__iter__().