New only_factor parameter for simulate_factors()
Add a new optional `only_factor` parameter for the simulate_factors()
method.

The only_factor parameter is specified, only a single simulation will be
run. The simulation that runs is for the factor at the index specified
by the value of the only_factor argument.

For example, if the snowbird command-line arguments look like this:

	meta.cmd_line: sb -f host.queue_depth 4,16 -f fm.fim.xfer_rate_MBps 400,800 -f fm.ase.t_wait_rd 4,5,6 -s duration 10 ms --only-factor 4

Then snowbird will pass in only_factor=4 as the argument to desmod and
the meta.sim.index and meta.sim.workspace will be set to 4 and
workspace/4 respectively.

Additionally, meta.sim.special will look like this:

    meta.sim.special:
    - [host.queue_depth, 4]
    - [fm.fim.xfer_rate_MBps, 800]
    - [fm.ase.t_wait_rd, 5]

That is, only a single simulation, for the factor indicated by
meta.sim.special, was run.

One use for this new feature is orchestrating the execution of multiple,
per-factor, simulations on separate machines (e.g., in the cloud) with a
single command (with multiple factors).