Repair silent truncation of config override
Previously when apply_user_config was called with a float value
in user_config to be applied to a key with an int value in config,
the float value would be coerced to int potentially resulting in
truncation.

For the special case of a new float value for a config key with an int
value, the float value is now assigned directly to the config key so
no truncation occurs.