Queue at-most and at-least events
Use new QueueWhenAtMostEvent and QueueWhenAtLeastEvent as building blocks
for QueueWhenFullEvent, QueueWhenAnyEvent, QueueWhenNotFullEvent, and
QueueWhenEmptyEvent (Queue.when_full(), Queue.when_any(),
Queue.when_not_full() and Queue.when_empty()).

This reduces the number of waiter lists that Queue has to maintain and
allows the user to wait for the Queue size to reach arbitrary levels with
when_at_least() and when_at_most().

Unit tests are documentation are updated for this change as well.

Addresses #18.