Repair Pool when_not_full and when_any methods
The when_not_full method creates a PoolWhenAtMostEvent based on an
epsilon distance from full. For an int valued Pool, a value of 0.5 is
sufficient and is supplied by default. For a float valued Pool, a working
value of epsilon depends on the Pool capacity and so the epsilon is a required
argument of when_not_full for a float valued Pool.

Similarly the when_any method creates a PoolWhenAtLeastEvent with an epsilon
distance from 0. The default epsilon is corrected from float_info.epsilon to
float_info.min.